[refactor] Move storage related functions to a separate module
